Render within internal app viewer when available
Note that this requires the master branch of shiny from github (for the ability to provide a function as the launch.browser parameter).

The RStudio internal application viewer is enabled in RStudio >= 0.98.414 (if an earlier version of RStudio is in use then it will gracefully fall back to using utils::browseURL). Note that for Shiny applications to work correctly inside RStudio the master branch of httpuv should be installed from github.