Prevent shiny-ggvis.js from throwing when loaded from non-Shiny app
knit_print.ggvis always seems to cause shiny-ggvis.js to get loaded,
even if used in a non-Shiny context. This change allows shiny-ggvis.js
to be safely loaded no matter what the context. The alternative would
be to be more careful about when shiny-ggvis.js is loaded; I couldn't
immediately figure out a good way to do that. It seems weird that
the static rendering codepaths should use ggvisOutput, but I assume
that was done (by me??) for a reason.